requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai==0.3.2
huggingface-hub==0.19.4
python-dotenv==1.0.0
marshmallow==3.20.1